don't let nwqid become negative
at the start of the loop nwqid is 0, so if the first component can't be
opened nwqid becomes -1 and since it's not 0, we end up calling np_walk
with -1 as length.  This in turns converts it back to uint16_t and we
generate an invalid packet.

The solution is to not decrement nwqid at all, it fixes all the current
tests case and is the correct behaviour that the rest of the code
expects.