implement virtual and userdata tables for users handling
The idea is to abstract over the users that kamid manages.  The virtual
table allows to use a custom userbase instead of the local UNIX users.
The userdata tables maps (virtual) users to their home directory, which
is the only directory tree exported by kamid as of now.

Together, they allow a powerful and flexible management of the users and
their data.  The idea is somewhat stolen from smtpd(8).