ensure that a walk with zero paths don't replies with a qid
A walk of zero paths is used to duplicate a fid.  it doesn't make sense
to reply with a body containing the qid relative to the file, because
it's not a real walk, it's just a duplication.  Furthermore, it would
break the expectations that the number of qid returned is less or equal
to the walked components.