comment out the 9P2000.L stuff
Initially I thought to support the .L dialect, but now I'm not that
sure.  Tread on a directory is not that hard, and I personally like that
more than the .L alternative of the Treaddir, it's ugly.  Some .L stuff
may be interesting to support, at least for symlinks, but at least for a
first version I want to keep it simple and implement plain 9P2000.