clarify qid implementation: they are only wrappers around direrctory fds
The bug recently found by cage is caused by twalk accidentaly reusing
the file descriptor for the newly created qid from the "parent" qid (the
start of the walk.)

I'm considering qids as wrappers around file descriptors.  This way, all
the 9p actions are straigtforward to implement using the *at functions
(openat, fstatat, unlinkat, ...) with only one path component.  The
mistake was to save the path component (a file name) inside the qid
struct: this leads to confusion and prevent us from sharing qids to our
heart's content.

The best way to fix the problem is to correct the mistake done in the
qid implementation: move the path name to the fid struct.  This needs
some fixes in various places.

One important difference is that now we're completely leveraging the
kernel for checking that clients don't open directory for writing.

All tests are now passing; many many thanks to cage for the "extra"
regression suite, it helped a lot in squashing the remaining bugs.