dis-entagle qids from the directory file descriptor
I've been using the qid both as the 9p protocol element *and* as a
wrapper with refcounting around a directory file descriptor.

I can do better, let's split the qid struct into a "proper" qid and dir.
Dir is now the wrapper with refcounting around the directory file
descriptor and qid is just the 9p struct.  fids are still associated
with a dir, but now they store the directory name if they point to the
directory.  The previous pattern to detect if a fid is pointing to a dir
thus changed: instead of assuming that fpath is "", look at the qid
struct.