reply with an Rerror if Rread/Rstat would overflow msize
Other replies have either a fixed reply size or a (relatively small)
maximum reply size.  Rread and Rstat on the other hand are easy to abuse
and generate replies longer that msize.

I don't care if a client manage to negotiate a ridiculy small msize
(maybe will add a minimum value in tversion in the future), but I feel
that Rread and Rstat need a small safety net.