fix cd
Walk can be successful in various conditions where cd has to fail.
For instance, we may have reached a file or we couldn't walk all the
required components.  In these "soft-fail" cases, the walk has mutated
the fids, so to abort the cd we have to do a walk using a second fid
and then clunk the old or new one depending on if it was _completely_
successful or not.

Another option may be to track the number of walked path and attempt
to walk '..' the right amount of times, this would saves a Tclunk in
the case the walk succeeded and require a slightly bigger 9P packet to
be sent in the failure case.  It would complicate the walk_path API
though, so I've not chosen this option.