make sure we don't draw a bar larger than the screen
if p->done goes up, way over p->max, we end up with perc being > 100.
in this case, reporting the correct percentage is fine, but drawing a
bar larger than the screen is not.

The bar could still "overflow" and occupy a second line, but only if
the percentage number takes more then 4 digits, so it's unlikely.