prepend builddir/kamid to PATH before running the lisp regress suite
I'm not proud of that hack, which is a double hack in fact, but I've
already wasted too much time on this.  The first hack is the PATH
mangling, and the second the $$(cd ... ; pwd) to obtain an *absolute*
path: for build in-tree top_builddir is just `.' which isn't really
helpful after a cd.