kamid: add wstat
it's not exactly compliant, as 9p requires the changes to be done
atomically, but it's not possible to better targeting a UNIX file
system.

The 9p manpage regarding wstat isn't clear about the behaviour when
changing the name of the file: kamid changes the associated fid to
reflect the rename.  kamid also allows to rename files to other
directories, the fid is updated to the new location even in this case.