fail if the received packet is too long
This adds extra checks to every t* function: if the client sent too
much data per message, treat is an a protocol violation and terminate
the connection.  Regress suite still passes, this is just an extra
safety check for finding wrong clients implementations.

The twrite function was the only one with a check like this.