wstat: missing read for size
9p stat message have a leading 2 bytes long size field.  It's meant to
simplify the parsing, because while reading the contents of a
directory it's necessary to know how much long an entry is, so the
length field is present even in the Rstat reply and in the Twstat
input.

Previously we didn't consume that bit and thus mis-read all of the
following fields.