ftp: truncate the file in woc_file
The idea behind woc_file is to create or open a file, but it's always
used for writing fresh new content, so it's actually an error that it
opens a file for writing without truncating it too.  It means that if we
write less content than what the file already has, we kepp garbage at
the end.

This adds an extra argument in send_fid (that is actually used only from
woc_file now) so that we can pass KOTRUNC for do_open.