kamiftp: automatic cd or fetch on the given path
before the given path was used for Tattach.  Turns out at least u9fs
doesn't seem to use that field for the initial directory (not sure if it
should).  kamiftp now always issues a Tattach with aname="/" and then
does a Twalk on the path: if it names a directory, it becomes the remote
working directory like a `cd' was issued, if it names a file it is
fetched and kamiftp quits.