kamiftp: don't walk dot
Since d3e1ab0 kamiftp started to issue Twalks over ".".  walk(9p) (from
plan9ports) says that

	The name ``.'' (dot), meaning the current directory, is not
	used in the protocol.

hinting that it shouldn't be used, and thus kamid rejects it.  It's also
redundant since it's always possible to do a walk to duplicate the
directory fid instead of walkig "." which is what kamiftp now does.